+++ Source/JNA/waffle-tomcat85/src/main/java/waffle/apache/MixedAuthenticator.java
@@ -1,7 +1,7 @@
 /**
  * Waffle (https://github.com/Waffle/waffle)
  *
- * Copyright (c) 2010-2018 Application Security, Inc.
+ * Copyright (c) 2010-2019 Application Security, Inc.
  *
  * All rights reserved. This program and the accompanying materials are made available under the terms of the Eclipse
  * Public License v1.0 which accompanies this distribution, and is available at
@@ -100,13 +100,7 @@ public class MixedAuthenticator extends WaffleAuthenticatorBase {
             this.sendUnauthorized(response);
             return false;
         } else if (securityCheck) {
-            final boolean postResult = this.post(request, response);
-            if (postResult) {
-                this.redirectTo(request, response, request.getServletPath());
-            } else {
-                this.redirectTo(request, response, loginConfig.getErrorPage());
-            }
-            return postResult;
+            return this.post(request, response);
         } else {
             this.redirectTo(request, response, loginConfig.getLoginPage());
             return false;